Splitted SeriesSonify composition from ChartSonify.
