+++ ts/Extensions/Sonification/Options.ts
@@ -14,18 +14,28 @@
 
 /* *
  *
+ * Imports
+ *
+ * */
+
+import type ChartSonify from './ChartSonify';
+import type SeriesSonify from './SeriesSonify';
+
+/* *
+ *
  * Declarations
  *
  * */
+
 declare module '../../Core/Options'{
     interface Options {
-        sonification?: Highcharts.ChartSonificationOptions;
+        sonification?: ChartSonify.ChartSonificationOptions;
     }
 }
 
 declare module '../../Core/Series/SeriesOptions' {
     interface SeriesOptions {
-        sonification?: Highcharts.SeriesSonificationOptions;
+        sonification?: SeriesSonify.SeriesSonificationOptions;
     }
 }
